requirements.txt
flask
Flask-Compress
Flask-Mail
Flask-WTF
supabase
python-dotenv
bcrypt
requests
